Split ext_proc/filter_test into smaller suites
