+++ test/extensions/filters/http/ext_proc/filter_test.cc
@@ -21,6 +21,7 @@
 
 #include "test/common/http/common.h"
 #include "test/common/http/conn_manager_impl_test_base.h"
+#include "test/extensions/filters/http/ext_proc/filter_test_common.h"
 #include "test/extensions/filters/http/ext_proc/mock_server.h"
 #include "test/extensions/filters/http/ext_proc/utils.h"
 #include "test/mocks/event/mocks.h"
@@ -79,589 +80,12 @@ using ::Envoy::Http::TestResponseTrailerMapImpl;
 using ::Envoy::Http::ExternalProcessing::SaveProcessingResponseFactory;
 using ::Envoy::Http::ExternalProcessing::SaveProcessingResponseFilterState;
 
-using ::testing::AnyNumber;
 using ::testing::Eq;
 using ::testing::Invoke;
 using ::testing::Return;
-using ::testing::ReturnRef;
 using ::testing::Unused;
 
 using namespace std::chrono_literals;
-
-static const uint32_t BufferSize = 100000;
-static const std::string filter_config_name = "scooby.dooby.doo";
-
-// These tests are all unit tests that directly drive an instance of the
-// ext_proc filter and verify the behavior using mocks.
-
-class HttpFilterTest : public testing::Test {
-protected:
-  void initialize(std::string&& yaml, bool is_upstream_filter = false) {
-    client_ = std::make_unique<MockClient>();
-    route_ = std::make_shared<NiceMock<Router::MockRoute>>();
-    EXPECT_CALL(*client_, start(_, _, _, _)).WillOnce(Invoke(this, &HttpFilterTest::doStart));
-    EXPECT_CALL(encoder_callbacks_, dispatcher()).WillRepeatedly(ReturnRef(dispatcher_));
-    EXPECT_CALL(decoder_callbacks_, dispatcher()).WillRepeatedly(ReturnRef(dispatcher_));
-    EXPECT_CALL(decoder_callbacks_, route()).WillRepeatedly(Return(route_));
-    EXPECT_CALL(decoder_callbacks_, streamInfo()).WillRepeatedly(ReturnRef(stream_info_));
-    EXPECT_CALL(encoder_callbacks_, streamInfo()).WillRepeatedly(ReturnRef(stream_info_));
-    EXPECT_CALL(stream_info_, dynamicMetadata()).WillRepeatedly(ReturnRef(dynamic_metadata_));
-    EXPECT_CALL(stream_info_, setDynamicMetadata(_, _))
-        .Times(AnyNumber())
-        .WillRepeatedly(Invoke(this, &HttpFilterTest::doSetDynamicMetadata));
-
-    EXPECT_CALL(decoder_callbacks_, connection())
-        .WillRepeatedly(Return(OptRef<const Network::Connection>{connection_}));
-    EXPECT_CALL(encoder_callbacks_, connection())
-        .WillRepeatedly(Return(OptRef<const Network::Connection>{connection_}));
-
-    // Pointing dispatcher_.time_system_ to a SimulatedTimeSystem object.
-    test_time_ = new Envoy::Event::SimulatedTimeSystem();
-    dispatcher_.time_system_.reset(test_time_);
-
-    EXPECT_CALL(dispatcher_, createTimer_(_))
-        .Times(AnyNumber())
-        .WillRepeatedly(Invoke([this](Unused) {
-          // Create a mock timer that we can check at destruction time to see if
-          // all timers were disabled no matter what. MockTimer has default
-          // actions that we just have to enable properly here.
-          auto* timer = new Event::MockTimer();
-          EXPECT_CALL(*timer, enableTimer(_, _)).Times(AnyNumber());
-          EXPECT_CALL(*timer, disableTimer()).Times(AnyNumber());
-          EXPECT_CALL(*timer, enabled()).Times(AnyNumber());
-          timers_.push_back(timer);
-          return timer;
-        }));
-    EXPECT_CALL(decoder_callbacks_, filterConfigName()).WillRepeatedly(Return(filter_config_name));
-
-    envoy::extensions::filters::http::ext_proc::v3::ExternalProcessor proto_config{};
-    if (!yaml.empty()) {
-      TestUtility::loadFromYaml(yaml, proto_config);
-    }
-    config_ = std::make_shared<FilterConfig>(
-        proto_config, 200ms, 10000, *stats_store_.rootScope(), "", is_upstream_filter,
-        Envoy::Extensions::Filters::Common::Expr::createBuilder(nullptr), factory_context_);
-    filter_ = std::make_unique<Filter>(config_, std::move(client_));
-    filter_->setEncoderFilterCallbacks(encoder_callbacks_);
-    EXPECT_CALL(encoder_callbacks_, encoderBufferLimit()).WillRepeatedly(Return(BufferSize));
-    filter_->setDecoderFilterCallbacks(decoder_callbacks_);
-    EXPECT_CALL(decoder_callbacks_, decoderBufferLimit()).WillRepeatedly(Return(BufferSize));
-    HttpTestUtility::addDefaultHeaders(request_headers_);
-    request_headers_.setMethod("POST");
-  }
-
-  void initializeTestSendAll() {
-    initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  processing_mode:
-    request_header_mode: "SEND"
-    response_header_mode: "SEND"
-    request_body_mode: "STREAMED"
-    response_body_mode: "STREAMED"
-    request_trailer_mode: "SEND"
-    response_trailer_mode: "SEND"
-  )EOF");
-  }
-
-  void TearDown() override {
-    // This will fail if, at the end of the test, we left any timers enabled.
-    // (This particular test suite does not actually let timers expire,
-    // although other test suites do.)
-    EXPECT_TRUE(allTimersDisabled());
-  }
-
-  bool allTimersDisabled() {
-    for (auto* t : timers_) {
-      if (t->enabled_) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  ExternalProcessorStreamPtr doStart(ExternalProcessorCallbacks& callbacks,
-                                     const Grpc::GrpcServiceConfigWithHashKey& config_with_hash_key,
-                                     const Envoy::Http::AsyncClient::StreamOptions&,
-                                     Envoy::Http::StreamFilterSidestreamWatermarkCallbacks&) {
-    if (final_expected_grpc_service_.has_value()) {
-      EXPECT_TRUE(TestUtility::protoEqual(final_expected_grpc_service_.value(),
-                                          config_with_hash_key.config()));
-    }
-
-    stream_callbacks_ = &callbacks;
-
-    auto stream = std::make_unique<NiceMock<MockStream>>();
-    // We never send with the "close" flag set
-    EXPECT_CALL(*stream, send(_, false)).WillRepeatedly(Invoke(this, &HttpFilterTest::doSend));
-
-    EXPECT_CALL(*stream, streamInfo()).WillRepeatedly(ReturnRef(async_client_stream_info_));
-
-    // close is idempotent and only called once per filter
-    EXPECT_CALL(*stream, close()).WillOnce(Invoke(this, &HttpFilterTest::doSendClose));
-
-    return stream;
-  }
-
-  void doSetDynamicMetadata(const std::string& ns, const Protobuf::Struct& val) {
-    (*dynamic_metadata_.mutable_filter_metadata())[ns] = val;
-  };
-
-  void doSend(ProcessingRequest&& request, Unused) { last_request_ = std::move(request); }
-
-  bool doSendClose() { return !server_closed_stream_; }
-
-  void setUpDecodingBuffering(Buffer::Instance& buf, bool expect_modification = false) {
-    EXPECT_CALL(decoder_callbacks_, decodingBuffer()).WillRepeatedly(Return(&buf));
-    EXPECT_CALL(decoder_callbacks_, addDecodedData(_, false))
-        .WillRepeatedly(
-            Invoke([&buf](Buffer::Instance& new_chunk, Unused) { buf.add(new_chunk); }));
-    if (expect_modification) {
-      EXPECT_CALL(decoder_callbacks_, modifyDecodingBuffer(_))
-          .WillOnce(
-              Invoke([&buf](std::function<void(Buffer::Instance&)> callback) { callback(buf); }));
-    }
-  }
-
-  void setUpEncodingBuffering(Buffer::Instance& buf, bool expect_modification = false) {
-    EXPECT_CALL(encoder_callbacks_, encodingBuffer()).WillRepeatedly(Return(&buf));
-    EXPECT_CALL(encoder_callbacks_, addEncodedData(_, false))
-        .WillRepeatedly(
-            Invoke([&buf](Buffer::Instance& new_chunk, Unused) { buf.add(new_chunk); }));
-    if (expect_modification) {
-      EXPECT_CALL(encoder_callbacks_, modifyEncodingBuffer(_))
-          .WillOnce(
-              Invoke([&buf](std::function<void(Buffer::Instance&)> callback) { callback(buf); }));
-    }
-  }
-
-  void setUpDecodingWatermarking(bool& watermarked) {
-    EXPECT_CALL(decoder_callbacks_, onDecoderFilterAboveWriteBufferHighWatermark())
-        .WillRepeatedly(Invoke([&watermarked]() {
-          EXPECT_FALSE(watermarked);
-          watermarked = true;
-        }));
-    EXPECT_CALL(decoder_callbacks_, onDecoderFilterBelowWriteBufferLowWatermark())
-        .WillRepeatedly(Invoke([&watermarked]() {
-          EXPECT_TRUE(watermarked);
-          watermarked = false;
-        }));
-  }
-
-  void setUpEncodingWatermarking(bool& watermarked) {
-    EXPECT_CALL(encoder_callbacks_, onEncoderFilterAboveWriteBufferHighWatermark())
-        .WillRepeatedly(Invoke([&watermarked]() {
-          EXPECT_FALSE(watermarked);
-          watermarked = true;
-        }));
-    EXPECT_CALL(encoder_callbacks_, onEncoderFilterBelowWriteBufferLowWatermark())
-        .WillRepeatedly(Invoke([&watermarked]() {
-          EXPECT_TRUE(watermarked);
-          watermarked = false;
-        }));
-  }
-
-  // Expect a request_headers request, and send back a valid response.
-  void processRequestHeaders(
-      bool buffering_data,
-      absl::optional<std::function<void(const HttpHeaders&, ProcessingResponse&, HeadersResponse&)>>
-          cb) {
-    ASSERT_TRUE(last_request_.has_request_headers());
-    const auto& headers = last_request_.request_headers();
-
-    auto response = std::make_unique<ProcessingResponse>();
-    auto* headers_response = response->mutable_request_headers();
-    if (cb) {
-      (*cb)(headers, *response, *headers_response);
-    }
-
-    if (observability_mode_) {
-      EXPECT_TRUE(last_request_.observability_mode());
-      return;
-    }
-
-    EXPECT_FALSE(last_request_.observability_mode());
-    if (!buffering_data) {
-      EXPECT_CALL(decoder_callbacks_, continueDecoding());
-    }
-    test_time_->advanceTimeWait(std::chrono::microseconds(10));
-    stream_callbacks_->onReceiveMessage(std::move(response));
-  }
-
-  // Expect a response_headers request, and send back a valid response
-  void processResponseHeaders(
-      bool buffering_data,
-      absl::optional<std::function<void(const HttpHeaders&, ProcessingResponse&, HeadersResponse&)>>
-          cb) {
-    ASSERT_TRUE(last_request_.has_response_headers());
-    const auto& headers = last_request_.response_headers();
-    auto response = std::make_unique<ProcessingResponse>();
-    auto* headers_response = response->mutable_response_headers();
-    if (cb) {
-      (*cb)(headers, *response, *headers_response);
-    }
-
-    if (observability_mode_) {
-      EXPECT_TRUE(last_request_.observability_mode());
-      return;
-    }
-
-    EXPECT_FALSE(last_request_.observability_mode());
-    if (!buffering_data) {
-      EXPECT_CALL(encoder_callbacks_, continueEncoding());
-    }
-    test_time_->advanceTimeWait(std::chrono::microseconds(10));
-    stream_callbacks_->onReceiveMessage(std::move(response));
-  }
-
-  void processResponseHeadersAfterTrailer(
-      absl::optional<std::function<void(const HttpHeaders&, ProcessingResponse&, HeadersResponse&)>>
-          cb) {
-    HttpHeaders headers;
-    auto response = std::make_unique<ProcessingResponse>();
-    auto* headers_response = response->mutable_response_headers();
-    if (cb) {
-      (*cb)(headers, *response, *headers_response);
-    }
-    test_time_->advanceTimeWait(std::chrono::microseconds(10));
-    stream_callbacks_->onReceiveMessage(std::move(response));
-  }
-
-  // Expect a request_body request, and send back a valid response
-  void processRequestBody(
-      absl::optional<std::function<void(const HttpBody&, ProcessingResponse&, BodyResponse&)>> cb,
-      bool should_continue = true,
-      const std::chrono::microseconds latency = std::chrono::microseconds(10)) {
-    ASSERT_TRUE(last_request_.has_request_body());
-
-    if (observability_mode_) {
-      EXPECT_TRUE(last_request_.observability_mode());
-      return;
-    }
-
-    EXPECT_FALSE(last_request_.observability_mode());
-    const auto& body = last_request_.request_body();
-    auto response = std::make_unique<ProcessingResponse>();
-    auto* body_response = response->mutable_request_body();
-    if (cb) {
-      (*cb)(body, *response, *body_response);
-    }
-    if (should_continue) {
-      EXPECT_CALL(decoder_callbacks_, continueDecoding());
-    }
-    test_time_->advanceTimeWait(latency);
-    stream_callbacks_->onReceiveMessage(std::move(response));
-  }
-
-  // Expect a request_body request, and send back a valid response
-  void processResponseBody(
-      absl::optional<std::function<void(const HttpBody&, ProcessingResponse&, BodyResponse&)>> cb,
-      bool should_continue = true) {
-    ASSERT_TRUE(last_request_.has_response_body());
-
-    if (observability_mode_) {
-      EXPECT_TRUE(last_request_.observability_mode());
-      return;
-    }
-
-    EXPECT_FALSE(last_request_.observability_mode());
-    const auto& body = last_request_.response_body();
-    auto response = std::make_unique<ProcessingResponse>();
-    auto* body_response = response->mutable_response_body();
-    if (cb) {
-      (*cb)(body, *response, *body_response);
-    }
-    if (should_continue) {
-      EXPECT_CALL(encoder_callbacks_, continueEncoding());
-    }
-    test_time_->advanceTimeWait(std::chrono::microseconds(10));
-    stream_callbacks_->onReceiveMessage(std::move(response));
-  }
-
-  void processResponseBodyHelper(absl::string_view data, Buffer::OwnedImpl& want_response_body,
-                                 bool end_of_stream = false, bool should_continue = false) {
-    processResponseBody(
-        [&](const HttpBody&, ProcessingResponse&, BodyResponse& resp) {
-          auto* streamed_response =
-              resp.mutable_response()->mutable_body_mutation()->mutable_streamed_response();
-          streamed_response->set_end_of_stream(end_of_stream);
-          streamed_response->set_body(data);
-          want_response_body.add(data);
-        },
-        should_continue);
-  }
-
-  void processResponseBodyStreamedAfterTrailer(absl::string_view data,
-                                               Buffer::OwnedImpl& want_response_body) {
-    auto response = std::make_unique<ProcessingResponse>();
-    auto* body_response = response->mutable_response_body();
-    auto* streamed_response =
-        body_response->mutable_response()->mutable_body_mutation()->mutable_streamed_response();
-    streamed_response->set_body(data);
-    want_response_body.add(data);
-    test_time_->advanceTimeWait(std::chrono::microseconds(10));
-    stream_callbacks_->onReceiveMessage(std::move(response));
-  }
-
-  void processRequestTrailers(
-      absl::optional<
-          std::function<void(const HttpTrailers&, ProcessingResponse&, TrailersResponse&)>>
-          cb,
-      bool should_continue = true) {
-    ASSERT_TRUE(last_request_.has_request_trailers());
-
-    if (observability_mode_) {
-      EXPECT_TRUE(last_request_.observability_mode());
-      return;
-    }
-
-    EXPECT_FALSE(last_request_.observability_mode());
-    const auto& trailers = last_request_.request_trailers();
-    auto response = std::make_unique<ProcessingResponse>();
-    auto* trailers_response = response->mutable_request_trailers();
-    if (cb) {
-      (*cb)(trailers, *response, *trailers_response);
-    }
-    if (should_continue) {
-      EXPECT_CALL(decoder_callbacks_, continueDecoding());
-    }
-    test_time_->advanceTimeWait(std::chrono::microseconds(10));
-    stream_callbacks_->onReceiveMessage(std::move(response));
-  }
-
-  void processResponseTrailers(
-      absl::optional<
-          std::function<void(const HttpTrailers&, ProcessingResponse&, TrailersResponse&)>>
-          cb,
-      bool should_continue = true) {
-    ASSERT_TRUE(last_request_.has_response_trailers());
-    if (observability_mode_) {
-      EXPECT_TRUE(last_request_.observability_mode());
-      return;
-    }
-    EXPECT_FALSE(last_request_.observability_mode());
-
-    const auto& trailers = last_request_.response_trailers();
-    auto response = std::make_unique<ProcessingResponse>();
-    auto* trailers_response = response->mutable_response_trailers();
-    if (cb) {
-      (*cb)(trailers, *response, *trailers_response);
-    }
-    if (should_continue) {
-      EXPECT_CALL(encoder_callbacks_, continueEncoding());
-    }
-    test_time_->advanceTimeWait(std::chrono::microseconds(10));
-    stream_callbacks_->onReceiveMessage(std::move(response));
-  }
-
-  // Get the gRPC call stats data from the filter state.
-  const ExtProcLoggingInfo::GrpcCalls&
-  getGrpcCalls(const envoy::config::core::v3::TrafficDirection traffic_direction) {
-    // The number of processor grpc calls made in the encoding and decoding path.
-    const ExtProcLoggingInfo::GrpcCalls& grpc_calls =
-        stream_info_.filterState()
-            ->getDataReadOnly<
-                Envoy::Extensions::HttpFilters::ExternalProcessing::ExtProcLoggingInfo>(
-                filter_config_name)
-            ->grpcCalls(traffic_direction);
-    return grpc_calls;
-  }
-
-  // Check gRPC call stats for headers and trailers.
-  void checkGrpcCall(const ExtProcLoggingInfo::GrpcCall call,
-                     const std::chrono::microseconds latency,
-                     const Grpc::Status::GrpcStatus call_status) {
-    EXPECT_TRUE(call.latency_ == latency);
-    EXPECT_TRUE(call.call_status_ == call_status);
-  }
-
-  // Check gRPC call stats for body.
-  void checkGrpcCallBody(const ExtProcLoggingInfo::GrpcCallBody call, const uint32_t call_count,
-                         const Grpc::Status::GrpcStatus call_status,
-                         const std::chrono::microseconds total_latency,
-                         const std::chrono::microseconds max_latency,
-                         const std::chrono::microseconds min_latency) {
-    EXPECT_TRUE(call.call_count_ == call_count);
-    EXPECT_TRUE(call.last_call_status_ == call_status);
-    EXPECT_TRUE(call.total_latency_ == total_latency);
-    EXPECT_TRUE(call.max_latency_ == max_latency);
-    EXPECT_TRUE(call.min_latency_ == min_latency);
-  }
-
-  // Verify gRPC calls only happened on headers.
-  void
-  checkGrpcCallHeaderOnlyStats(const envoy::config::core::v3::TrafficDirection traffic_direction,
-                               const Grpc::Status::GrpcStatus call_status = Grpc::Status::Ok) {
-    auto& grpc_calls = getGrpcCalls(traffic_direction);
-    EXPECT_TRUE(grpc_calls.header_stats_ != nullptr);
-    checkGrpcCall(*grpc_calls.header_stats_, std::chrono::microseconds(10), call_status);
-    EXPECT_TRUE(grpc_calls.trailer_stats_ == nullptr);
-    EXPECT_TRUE(grpc_calls.body_stats_ == nullptr);
-  }
-
-  // Verify gRPC calls for  headers, body, and trailer.
-  void checkGrpcCallStatsAll(const envoy::config::core::v3::TrafficDirection traffic_direction,
-                             const uint32_t body_chunk_number,
-                             const Grpc::Status::GrpcStatus body_call_status = Grpc::Status::Ok,
-                             const bool trailer_stats = true) {
-    auto& grpc_calls = getGrpcCalls(traffic_direction);
-    EXPECT_TRUE(grpc_calls.header_stats_ != nullptr);
-    checkGrpcCall(*grpc_calls.header_stats_, std::chrono::microseconds(10), Grpc::Status::Ok);
-
-    if (trailer_stats) {
-      EXPECT_TRUE(grpc_calls.trailer_stats_ != nullptr);
-      checkGrpcCall(*grpc_calls.trailer_stats_, std::chrono::microseconds(10), Grpc::Status::Ok);
-    } else {
-      EXPECT_TRUE(grpc_calls.trailer_stats_ == nullptr);
-    }
-
-    EXPECT_TRUE(grpc_calls.body_stats_ != nullptr);
-    checkGrpcCallBody(*grpc_calls.body_stats_, body_chunk_number, body_call_status,
-                      std::chrono::microseconds(10) * body_chunk_number,
-                      std::chrono::microseconds(10), std::chrono::microseconds(10));
-  }
-
-  // Verify no gRPC call happened.
-  void expectNoGrpcCall(const envoy::config::core::v3::TrafficDirection traffic_direction) {
-    auto& grpc_calls = getGrpcCalls(traffic_direction);
-    EXPECT_TRUE(grpc_calls.header_stats_ == nullptr);
-    EXPECT_TRUE(grpc_calls.trailer_stats_ == nullptr);
-    EXPECT_TRUE(grpc_calls.body_stats_ == nullptr);
-  }
-
-  void sendChunkRequestData(const uint32_t chunk_number, const bool send_grpc) {
-    for (uint32_t i = 0; i < chunk_number; i++) {
-      Buffer::OwnedImpl req_data("foo");
-      EXPECT_EQ(FilterDataStatus::Continue, filter_->decodeData(req_data, false));
-      if (send_grpc) {
-        processRequestBody(absl::nullopt, false);
-      }
-    }
-  }
-
-  void sendChunkResponseData(const uint32_t chunk_number, const bool send_grpc) {
-    for (uint32_t i = 0; i < chunk_number; i++) {
-      Buffer::OwnedImpl resp_data("bar");
-      EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(resp_data, false));
-      if (send_grpc) {
-        processResponseBody(absl::nullopt, false);
-      }
-    }
-  }
-
-  void streamingSmallChunksWithBodyMutation(bool empty_last_chunk, bool mutate_last_chunk) {
-    initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  processing_mode:
-    request_header_mode: "SKIP"
-    response_header_mode: "SKIP"
-    request_body_mode: "NONE"
-    response_body_mode: "STREAMED"
-    request_trailer_mode: "SKIP"
-    response_trailer_mode: "SKIP"
-  )EOF");
-
-    EXPECT_EQ(FilterHeadersStatus::Continue, filter_->decodeHeaders(request_headers_, false));
-
-    Buffer::OwnedImpl first_chunk("foo");
-    EXPECT_EQ(FilterDataStatus::Continue, filter_->decodeData(first_chunk, false));
-    EXPECT_EQ(FilterTrailersStatus::Continue, filter_->decodeTrailers(request_trailers_));
-
-    response_headers_.addCopy(LowerCaseString(":status"), "200");
-    response_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-    EXPECT_EQ(FilterHeadersStatus::Continue, filter_->encodeHeaders(response_headers_, false));
-
-    Buffer::OwnedImpl want_response_body;
-    Buffer::OwnedImpl got_response_body;
-    EXPECT_CALL(encoder_callbacks_, injectEncodedDataToFilterChain(_, _))
-        .WillRepeatedly(Invoke([&got_response_body](Buffer::Instance& data, Unused) {
-          got_response_body.move(data);
-        }));
-    uint32_t chunk_number = 3;
-    for (uint32_t i = 0; i < chunk_number; i++) {
-      Buffer::OwnedImpl resp_data(std::to_string(i));
-      EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(resp_data, false));
-      processResponseBody(
-          [i, &want_response_body](const HttpBody& body, ProcessingResponse&, BodyResponse& resp) {
-            auto* body_mut = resp.mutable_response()->mutable_body_mutation();
-            body_mut->set_body(body.body() + " " + std::to_string(i) + " ");
-            want_response_body.add(body.body() + " " + std::to_string(i) + " ");
-          },
-          false);
-    }
-
-    std::string last_chunk_str = "";
-    Buffer::OwnedImpl resp_data;
-    if (!empty_last_chunk) {
-      last_chunk_str = std::to_string(chunk_number);
-    }
-    resp_data.add(last_chunk_str);
-    EXPECT_EQ(FilterDataStatus::StopIterationNoBuffer, filter_->encodeData(resp_data, true));
-    if (mutate_last_chunk) {
-      processResponseBody(
-          [&chunk_number, &want_response_body](const HttpBody& body, ProcessingResponse&,
-                                               BodyResponse& resp) {
-            auto* body_mut = resp.mutable_response()->mutable_body_mutation();
-            body_mut->set_body(body.body() + " " + std::to_string(chunk_number) + " ");
-            want_response_body.add(body.body() + " " + std::to_string(chunk_number) + " ");
-          },
-          true);
-    } else {
-      processResponseBody(absl::nullopt, true);
-      want_response_body.add(last_chunk_str);
-    }
-
-    EXPECT_EQ(want_response_body.toString(), got_response_body.toString());
-    filter_->onDestroy();
-
-    EXPECT_EQ(1, config_->stats().streams_started_.value());
-    EXPECT_EQ(4, config_->stats().stream_msgs_sent_.value());
-    EXPECT_EQ(4, config_->stats().stream_msgs_received_.value());
-    EXPECT_EQ(1, config_->stats().streams_closed_.value());
-  }
-
-  // The metadata configured as part of ext_proc filter should be in the filter state.
-  // In addition, bytes sent/received should also be stored.
-  void expectFilterState(const Envoy::Protobuf::Struct& expected_metadata) {
-    const auto* filterState =
-        stream_info_.filterState()
-            ->getDataReadOnly<
-                Envoy::Extensions::HttpFilters::ExternalProcessing::ExtProcLoggingInfo>(
-                filter_config_name);
-    const Envoy::Protobuf::Struct& loggedMetadata = filterState->filterMetadata();
-    EXPECT_THAT(loggedMetadata, ProtoEq(expected_metadata));
-  }
-
-  absl::optional<envoy::config::core::v3::GrpcService> final_expected_grpc_service_;
-  Grpc::GrpcServiceConfigWithHashKey config_with_hash_key_;
-  std::unique_ptr<MockClient> client_;
-  ExternalProcessorCallbacks* stream_callbacks_ = nullptr;
-  ProcessingRequest last_request_;
-  bool server_closed_stream_ = false;
-  bool observability_mode_ = false;
-  testing::NiceMock<Stats::MockIsolatedStatsStore> stats_store_;
-  FilterConfigSharedPtr config_;
-  std::shared_ptr<Filter> filter_;
-  testing::NiceMock<Event::MockDispatcher> dispatcher_;
-  testing::NiceMock<::Envoy::Http::MockStreamDecoderFilterCallbacks> decoder_callbacks_;
-  testing::NiceMock<::Envoy::Http::MockStreamEncoderFilterCallbacks> encoder_callbacks_;
-  Router::RouteConstSharedPtr route_;
-  testing::NiceMock<StreamInfo::MockStreamInfo> stream_info_;
-  testing::NiceMock<StreamInfo::MockStreamInfo> async_client_stream_info_;
-  TestRequestHeaderMapImpl request_headers_;
-  TestResponseHeaderMapImpl response_headers_;
-  TestRequestTrailerMapImpl request_trailers_;
-  TestResponseTrailerMapImpl response_trailers_;
-  std::vector<Event::MockTimer*> timers_;
-  Event::MockTimer* deferred_close_timer_;
-  Envoy::Event::SimulatedTimeSystem* test_time_;
-  envoy::config::core::v3::Metadata dynamic_metadata_;
-  testing::NiceMock<Network::MockConnection> connection_;
-  NiceMock<Server::Configuration::MockServerFactoryContext> factory_context_;
-};
 
 // Using the default configuration, test the filter with a processor that
 // replies to the request_headers message with an empty response
@@ -1624,226 +1048,6 @@ TEST_F(HttpFilterTest, StreamingDataSmallChunk) {
 
   checkGrpcCallStatsAll(envoy::config::core::v3::TrafficDirection::INBOUND, chunk_number);
   checkGrpcCallStatsAll(envoy::config::core::v3::TrafficDirection::OUTBOUND, 2 * chunk_number);
-}
-
-// Test suite for failure_mode_allow override.
-class FailureModeAllowOverrideTest : public HttpFilterTest {};
-
-TEST_F(FailureModeAllowOverrideTest, FilterAllowRouteDisallow) {
-  std::string yaml_config = R"EOF(
-grpc_service:
-  envoy_grpc:
-    cluster_name: "ext_proc_server"
-failure_mode_allow: true)EOF";
-  initialize(std::move(yaml_config));
-
-  ExtProcPerRoute route_proto;
-  route_proto.mutable_overrides()->mutable_failure_mode_allow()->set_value(false);
-  FilterConfigPerRoute route_config(route_proto);
-  EXPECT_CALL(decoder_callbacks_, perFilterConfigs())
-      .WillRepeatedly(
-          testing::Invoke([&]() -> Router::RouteSpecificFilterConfigs { return {&route_config}; }));
-
-  EXPECT_EQ(filter_->decodeHeaders(request_headers_, false), FilterHeadersStatus::StopIteration);
-  test_time_->advanceTimeWait(std::chrono::microseconds(10));
-
-  TestResponseHeaderMapImpl immediate_response_headers;
-  EXPECT_CALL(encoder_callbacks_,
-              sendLocalReply(::Envoy::Http::Code::InternalServerError, "", _, Eq(absl::nullopt),
-                             "ext_proc_error_gRPC_error_13{error_message}"))
-      .WillOnce(Invoke([&immediate_response_headers](
-                           Unused, Unused,
-                           std::function<void(ResponseHeaderMap & headers)> modify_headers, Unused,
-                           Unused) { modify_headers(immediate_response_headers); }));
-
-  server_closed_stream_ = true; // Simulate stream close without proper gRPC response
-  stream_callbacks_->onGrpcError(Grpc::Status::Internal, "error_message");
-  filter_->onDestroy();
-}
-
-TEST_F(FailureModeAllowOverrideTest, FilterDisallowRouteAllow) {
-  std::string yaml_config = R"EOF(
-grpc_service:
-  envoy_grpc:
-    cluster_name: "ext_proc_server"
-failure_mode_allow: false)EOF";
-  initialize(std::move(yaml_config));
-
-  ExtProcPerRoute route_proto;
-  route_proto.mutable_overrides()->mutable_failure_mode_allow()->set_value(true);
-  FilterConfigPerRoute route_config(route_proto);
-  EXPECT_CALL(decoder_callbacks_, perFilterConfigs())
-      .WillRepeatedly(
-          testing::Invoke([&]() -> Router::RouteSpecificFilterConfigs { return {&route_config}; }));
-
-  EXPECT_EQ(filter_->decodeHeaders(request_headers_, false), FilterHeadersStatus::StopIteration);
-  test_time_->advanceTimeWait(std::chrono::microseconds(10));
-
-  EXPECT_CALL(decoder_callbacks_, continueDecoding());
-
-  server_closed_stream_ = true; // Simulate stream close without proper gRPC response
-  stream_callbacks_->onGrpcError(Grpc::Status::Internal, "error_message");
-
-  Buffer::OwnedImpl req_data("foo");
-  EXPECT_EQ(filter_->decodeData(req_data, true), FilterDataStatus::Continue);
-  EXPECT_EQ(filter_->decodeTrailers(request_trailers_), FilterTrailersStatus::Continue);
-  EXPECT_EQ(filter_->encodeHeaders(response_headers_, true), FilterHeadersStatus::Continue);
-  filter_->onDestroy();
-
-  EXPECT_EQ(config_->stats().failure_mode_allowed_.value(), 1);
-}
-
-TEST_F(FailureModeAllowOverrideTest, FilterAllowNoRouteOverride) {
-  std::string yaml_config = R"EOF(
-grpc_service:
-  envoy_grpc:
-    cluster_name: "ext_proc_server"
-failure_mode_allow: true)EOF";
-  initialize(std::move(yaml_config));
-
-  EXPECT_EQ(filter_->decodeHeaders(request_headers_, false), FilterHeadersStatus::StopIteration);
-  test_time_->advanceTimeWait(std::chrono::microseconds(10));
-
-  EXPECT_CALL(decoder_callbacks_, continueDecoding());
-
-  server_closed_stream_ = true; // Simulate stream close without proper gRPC response
-  stream_callbacks_->onGrpcError(Grpc::Status::Internal, "error_message");
-
-  Buffer::OwnedImpl req_data("foo");
-  EXPECT_EQ(filter_->decodeData(req_data, true), FilterDataStatus::Continue);
-  EXPECT_EQ(filter_->decodeTrailers(request_trailers_), FilterTrailersStatus::Continue);
-  EXPECT_EQ(filter_->encodeHeaders(response_headers_, true), FilterHeadersStatus::Continue);
-  filter_->onDestroy();
-
-  EXPECT_EQ(config_->stats().failure_mode_allowed_.value(), 1);
-}
-
-// Verify that when status_on_error is configured, gRPC errors use the configured HTTP status.
-TEST_F(HttpFilterTest, GrpcErrorUsesConfiguredStatusOnError) {
-  std::string yaml_config = R"EOF(
-grpc_service:
-  envoy_grpc:
-    cluster_name: "ext_proc_server"
-status_on_error:
-  code: 503
-)EOF";
-  initialize(std::move(yaml_config));
-
-  // Start header processing to open the stream and arm timers.
-  EXPECT_EQ(filter_->decodeHeaders(request_headers_, false), FilterHeadersStatus::StopIteration);
-  test_time_->advanceTimeWait(std::chrono::microseconds(10));
-
-  // Expect a 503 Service Unavailable local reply instead of the previous default 500.
-  TestResponseHeaderMapImpl immediate_response_headers;
-  EXPECT_CALL(encoder_callbacks_,
-              sendLocalReply(::Envoy::Http::Code::ServiceUnavailable, "", _, Eq(absl::nullopt),
-                             "ext_proc_error_gRPC_error_13{error_message}"))
-      .WillOnce(Invoke([&immediate_response_headers](
-                           Unused, Unused,
-                           std::function<void(ResponseHeaderMap & headers)> modify_headers, Unused,
-                           Unused) { modify_headers(immediate_response_headers); }));
-
-  // Simulate a gRPC error from the external processor.
-  server_closed_stream_ = true;
-  stream_callbacks_->onGrpcError(Grpc::Status::Internal, "error_message");
-  filter_->onDestroy();
-}
-
-TEST_F(FailureModeAllowOverrideTest, FilterDisallowNoRouteOverride) {
-  std::string yaml_config = R"EOF(
-grpc_service:
-  envoy_grpc:
-    cluster_name: "ext_proc_server"
-failure_mode_allow: false)EOF";
-  initialize(std::move(yaml_config));
-
-  EXPECT_EQ(filter_->decodeHeaders(request_headers_, false), FilterHeadersStatus::StopIteration);
-  test_time_->advanceTimeWait(std::chrono::microseconds(10));
-
-  TestResponseHeaderMapImpl immediate_response_headers;
-  EXPECT_CALL(encoder_callbacks_,
-              sendLocalReply(::Envoy::Http::Code::InternalServerError, "", _, Eq(absl::nullopt),
-                             "ext_proc_error_gRPC_error_13{error_message}"))
-      .WillOnce(Invoke([&immediate_response_headers](
-                           Unused, Unused,
-                           std::function<void(ResponseHeaderMap & headers)> modify_headers, Unused,
-                           Unused) { modify_headers(immediate_response_headers); }));
-
-  server_closed_stream_ = true; // Simulate stream close without proper gRPC response
-  stream_callbacks_->onGrpcError(Grpc::Status::Internal, "error_message");
-  filter_->onDestroy();
-}
-
-TEST_F(FailureModeAllowOverrideTest, FilterAllowRouteUnrelatedOverride) {
-  std::string yaml_config = R"EOF(
-grpc_service:
-  envoy_grpc:
-    cluster_name: "ext_proc_server"
-failure_mode_allow: true)EOF";
-  initialize(std::move(yaml_config));
-
-  ExtProcPerRoute route_proto;
-  // This override does not set failure_mode_allow, so the filter's value should still apply.
-  route_proto.mutable_overrides()->mutable_processing_mode()->set_response_header_mode(
-      ProcessingMode::SKIP);
-  FilterConfigPerRoute route_config(route_proto);
-  EXPECT_CALL(decoder_callbacks_, perFilterConfigs())
-      .WillRepeatedly(
-          testing::Invoke([&]() -> Router::RouteSpecificFilterConfigs { return {&route_config}; }));
-
-  EXPECT_EQ(filter_->decodeHeaders(request_headers_, false), FilterHeadersStatus::StopIteration);
-  test_time_->advanceTimeWait(std::chrono::microseconds(10));
-
-  EXPECT_CALL(decoder_callbacks_, continueDecoding());
-
-  server_closed_stream_ = true; // Simulate stream close without proper gRPC response
-  stream_callbacks_->onGrpcError(Grpc::Status::Internal, "error_message");
-
-  Buffer::OwnedImpl req_data("foo");
-  EXPECT_EQ(filter_->decodeData(req_data, true), FilterDataStatus::Continue);
-  EXPECT_EQ(filter_->decodeTrailers(request_trailers_), FilterTrailersStatus::Continue);
-  EXPECT_EQ(filter_->encodeHeaders(response_headers_, true), FilterHeadersStatus::Continue);
-  filter_->onDestroy();
-
-  EXPECT_EQ(config_->stats().failure_mode_allowed_.value(), 1);
-}
-
-TEST_F(FailureModeAllowOverrideTest, FailureModeAllowMergedOverride) {
-  std::string yaml_config = R"EOF(
-grpc_service:
-  envoy_grpc:
-    cluster_name: "ext_proc_server"
-failure_mode_allow: true)EOF";
-  initialize(std::move(yaml_config));
-
-  ExtProcPerRoute route_proto_less_specific;
-  route_proto_less_specific.mutable_overrides()->mutable_failure_mode_allow()->set_value(true);
-  FilterConfigPerRoute route_config_less_specific(route_proto_less_specific);
-
-  ExtProcPerRoute route_proto_more_specific;
-  route_proto_more_specific.mutable_overrides()->mutable_failure_mode_allow()->set_value(false);
-  FilterConfigPerRoute route_config_more_specific(route_proto_more_specific);
-
-  EXPECT_CALL(decoder_callbacks_, perFilterConfigs())
-      .WillRepeatedly(testing::Invoke([&]() -> Router::RouteSpecificFilterConfigs {
-        return {&route_config_less_specific, &route_config_more_specific};
-      }));
-
-  EXPECT_EQ(filter_->decodeHeaders(request_headers_, false), FilterHeadersStatus::StopIteration);
-  test_time_->advanceTimeWait(std::chrono::microseconds(10));
-
-  TestResponseHeaderMapImpl immediate_response_headers;
-  EXPECT_CALL(encoder_callbacks_,
-              sendLocalReply(::Envoy::Http::Code::InternalServerError, "", _, Eq(absl::nullopt),
-                             "ext_proc_error_gRPC_error_13{error_message}"))
-      .WillOnce(Invoke([&immediate_response_headers](
-                           Unused, Unused,
-                           std::function<void(ResponseHeaderMap & headers)> modify_headers, Unused,
-                           Unused) { modify_headers(immediate_response_headers); }));
-
-  server_closed_stream_ = true; // Simulate stream close without proper gRPC response
-  stream_callbacks_->onGrpcError(Grpc::Status::Internal, "error_message");
-  filter_->onDestroy();
 }
 
 TEST_F(HttpFilterTest, StreamingBodyMutateLastEmptyChunk) {
@@ -2890,58 +2094,6 @@ TEST_F(HttpFilterTest, ProcessingModeOverrideResponseHeaders) {
   EXPECT_EQ(1, config_->stats().stream_msgs_sent_.value());
   EXPECT_EQ(1, config_->stats().stream_msgs_received_.value());
   EXPECT_EQ(1, config_->stats().streams_closed_.value());
-}
-
-// Set allow_mode_override in filter config to be true.
-// Set request_body_mode: FULL_DUPLEX_STREAMED
-// In such case, the mode_override in the response will be ignored.
-TEST_F(HttpFilterTest, DisableResponseModeOverrideByStreamedBodyMode) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  processing_mode:
-    request_header_mode: "SEND"
-    response_header_mode: "SEND"
-    request_body_mode: "FULL_DUPLEX_STREAMED"
-    response_body_mode: "FULL_DUPLEX_STREAMED"
-    request_trailer_mode: "SEND"
-    response_trailer_mode: "SEND"
-  allow_mode_override: true
-  )EOF");
-
-  EXPECT_EQ(filter_->config().allowModeOverride(), true);
-  EXPECT_EQ(filter_->config().sendBodyWithoutWaitingForHeaderResponse(), false);
-  EXPECT_EQ(filter_->config().processingMode().response_header_mode(), ProcessingMode::SEND);
-  EXPECT_EQ(filter_->config().processingMode().response_body_mode(),
-            ProcessingMode::FULL_DUPLEX_STREAMED);
-  EXPECT_EQ(filter_->config().processingMode().request_body_mode(),
-            ProcessingMode::FULL_DUPLEX_STREAMED);
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->decodeHeaders(request_headers_, true));
-
-  // When ext_proc server sends back the request header response, it contains the
-  // mode_override for the response_header_mode to be SKIP.
-  processRequestHeaders(
-      false, [](const HttpHeaders&, ProcessingResponse& response, HeadersResponse&) {
-        response.mutable_mode_override()->set_response_header_mode(ProcessingMode::SKIP);
-      });
-
-  response_headers_.addCopy(LowerCaseString(":status"), "200");
-  response_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->encodeHeaders(response_headers_, true));
-
-  // Verify such mode_override is ignored. The response header is still sent to the ext_proc server.
-  processResponseHeaders(false, [](const HttpHeaders& header_resp, ProcessingResponse&,
-                                   HeadersResponse&) {
-    EXPECT_TRUE(header_resp.end_of_stream());
-    TestRequestHeaderMapImpl expected_response{{":status", "200"}, {"content-type", "text/plain"}};
-    EXPECT_THAT(header_resp.headers(), HeaderProtosEqual(expected_response));
-  });
-
-  TestRequestHeaderMapImpl final_expected_response{{":status", "200"},
-                                                   {"content-type", "text/plain"}};
-  EXPECT_THAT(&response_headers_, HeaderMapEqualIgnoreOrder(&final_expected_response));
-  filter_->onDestroy();
 }
 
 // Set allow_mode_override in filter config to be true.
@@ -4717,325 +3869,6 @@ TEST_F(HttpFilterTest, StreamedTestInBothDirection) {
   filter_->onDestroy();
 }
 
-TEST_F(HttpFilterTest, DuplexStreamedBodyProcessingTestNormal) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  processing_mode:
-    response_body_mode: "FULL_DUPLEX_STREAMED"
-    response_trailer_mode: "SEND"
-  )EOF");
-
-  // Create synthetic HTTP request
-  HttpTestUtility::addDefaultHeaders(request_headers_);
-  request_headers_.setMethod("POST");
-  request_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->decodeHeaders(request_headers_, false));
-  processRequestHeaders(false, absl::nullopt);
-
-  response_headers_.addCopy(LowerCaseString(":status"), "200");
-  response_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  bool encoding_watermarked = false;
-  setUpEncodingWatermarking(encoding_watermarked);
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->encodeHeaders(response_headers_, false));
-  processResponseHeaders(false, absl::nullopt);
-
-  Buffer::OwnedImpl want_response_body;
-  Buffer::OwnedImpl got_response_body;
-  EXPECT_CALL(encoder_callbacks_, injectEncodedDataToFilterChain(_, _))
-      .WillRepeatedly(Invoke(
-          [&got_response_body](Buffer::Instance& data, Unused) { got_response_body.move(data); }));
-
-  // Test 7x3 streaming.
-  for (int i = 0; i < 7; i++) {
-    // 7 request chunks are sent to the ext_proc server.
-    Buffer::OwnedImpl resp_chunk;
-    TestUtility::feedBufferWithRandomCharacters(resp_chunk, 100);
-    EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(resp_chunk, false));
-  }
-
-  processResponseBodyHelper(" AAAAA ", want_response_body);
-  processResponseBodyHelper(" BBBB ", want_response_body);
-  processResponseBodyHelper(" CCC ", want_response_body);
-
-  // The two buffers should match.
-  EXPECT_EQ(want_response_body.toString(), got_response_body.toString());
-  EXPECT_FALSE(encoding_watermarked);
-
-  // Now do 1:1 streaming for a few chunks.
-  for (int i = 0; i < 3; i++) {
-    Buffer::OwnedImpl resp_chunk;
-    TestUtility::feedBufferWithRandomCharacters(resp_chunk, 100);
-    EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(resp_chunk, false));
-    processResponseBodyHelper(std::to_string(i), want_response_body);
-  }
-
-  // The two buffers should match.
-  EXPECT_EQ(want_response_body.toString(), got_response_body.toString());
-  EXPECT_FALSE(encoding_watermarked);
-
-  // Now send another 10 chunks.
-  for (int i = 0; i < 10; i++) {
-    Buffer::OwnedImpl resp_chunk;
-    TestUtility::feedBufferWithRandomCharacters(resp_chunk, 10);
-    EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(resp_chunk, false));
-  }
-  // Send the last chunk.
-  Buffer::OwnedImpl last_resp_chunk;
-  TestUtility::feedBufferWithRandomCharacters(last_resp_chunk, 10);
-  EXPECT_EQ(FilterDataStatus::StopIterationNoBuffer, filter_->encodeData(last_resp_chunk, true));
-
-  processResponseBodyHelper(" EEEEEEE ", want_response_body);
-  processResponseBodyHelper(" F ", want_response_body);
-  processResponseBodyHelper(" GGGGGGGGG ", want_response_body);
-  processResponseBodyHelper(" HH ", want_response_body, true, true);
-
-  // The two buffers should match.
-  EXPECT_EQ(want_response_body.toString(), got_response_body.toString());
-  EXPECT_FALSE(encoding_watermarked);
-  EXPECT_EQ(config_->stats().spurious_msgs_received_.value(), 0);
-  filter_->onDestroy();
-}
-
-TEST_F(HttpFilterTest, DuplexStreamedBodyProcessingTestWithTrailer) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  processing_mode:
-    response_body_mode: "FULL_DUPLEX_STREAMED"
-    response_trailer_mode: "SEND"
-  )EOF");
-
-  // Create synthetic HTTP request
-  HttpTestUtility::addDefaultHeaders(request_headers_);
-  request_headers_.setMethod("POST");
-  request_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->decodeHeaders(request_headers_, false));
-  processRequestHeaders(false, absl::nullopt);
-
-  response_headers_.addCopy(LowerCaseString(":status"), "200");
-  response_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  bool encoding_watermarked = false;
-  setUpEncodingWatermarking(encoding_watermarked);
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->encodeHeaders(response_headers_, false));
-  // Server sending headers response without waiting for body.
-  processResponseHeaders(false, absl::nullopt);
-
-  Buffer::OwnedImpl want_response_body;
-  Buffer::OwnedImpl got_response_body;
-  EXPECT_CALL(encoder_callbacks_, injectEncodedDataToFilterChain(_, _))
-      .WillRepeatedly(Invoke(
-          [&got_response_body](Buffer::Instance& data, Unused) { got_response_body.move(data); }));
-
-  for (int i = 0; i < 7; i++) {
-    // 7 request chunks are sent to the ext_proc server.
-    Buffer::OwnedImpl resp_chunk;
-    TestUtility::feedBufferWithRandomCharacters(resp_chunk, 100);
-    EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(resp_chunk, false));
-  }
-
-  EXPECT_EQ(FilterTrailersStatus::StopIteration, filter_->encodeTrailers(response_trailers_));
-
-  processResponseBodyStreamedAfterTrailer(" AAAAA ", want_response_body);
-  processResponseBodyStreamedAfterTrailer(" BBBB ", want_response_body);
-  processResponseTrailers(absl::nullopt, true);
-
-  // The two buffers should match.
-  EXPECT_EQ(want_response_body.toString(), got_response_body.toString());
-  EXPECT_FALSE(encoding_watermarked);
-
-  EXPECT_EQ(config_->stats().spurious_msgs_received_.value(), 0);
-  filter_->onDestroy();
-}
-
-TEST_F(HttpFilterTest, DuplexStreamedBodyProcessingTestWithHeaderAndTrailer) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  processing_mode:
-    response_header_mode: "SEND"
-    response_body_mode: "FULL_DUPLEX_STREAMED"
-    response_trailer_mode: "SEND"
-  )EOF");
-
-  // Create synthetic HTTP request
-  HttpTestUtility::addDefaultHeaders(request_headers_);
-  request_headers_.setMethod("POST");
-  request_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->decodeHeaders(request_headers_, false));
-  EXPECT_TRUE(last_request_.has_protocol_config());
-  processRequestHeaders(false, absl::nullopt);
-
-  response_headers_.addCopy(LowerCaseString(":status"), "200");
-  response_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  // Server buffer header, body and trailer before sending header response.
-  bool encoding_watermarked = false;
-  setUpEncodingWatermarking(encoding_watermarked);
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->encodeHeaders(response_headers_, false));
-
-  Buffer::OwnedImpl want_response_body;
-  Buffer::OwnedImpl got_response_body;
-  EXPECT_CALL(encoder_callbacks_, injectEncodedDataToFilterChain(_, _))
-      .WillRepeatedly(Invoke(
-          [&got_response_body](Buffer::Instance& data, Unused) { got_response_body.move(data); }));
-
-  for (int i = 0; i < 7; i++) {
-    // 7 request chunks are sent to the ext_proc server.
-    Buffer::OwnedImpl resp_chunk;
-    TestUtility::feedBufferWithRandomCharacters(resp_chunk, 100);
-    EXPECT_EQ(FilterDataStatus::StopIterationNoBuffer, filter_->encodeData(resp_chunk, false));
-  }
-
-  EXPECT_EQ(FilterTrailersStatus::StopIteration, filter_->encodeTrailers(response_trailers_));
-
-  EXPECT_FALSE(last_request_.has_protocol_config());
-  // Server now sends back response.
-  processResponseHeadersAfterTrailer(absl::nullopt);
-  processResponseBodyStreamedAfterTrailer(" AAAAA ", want_response_body);
-  processResponseBodyStreamedAfterTrailer(" BBBB ", want_response_body);
-  processResponseTrailers(absl::nullopt, true);
-
-  // The two buffers should match.
-  EXPECT_EQ(want_response_body.toString(), got_response_body.toString());
-  EXPECT_FALSE(encoding_watermarked);
-
-  EXPECT_EQ(config_->stats().spurious_msgs_received_.value(), 0);
-  filter_->onDestroy();
-}
-
-TEST_F(HttpFilterTest, DuplexStreamedBodyProcessingTestWithHeaderAndTrailerNoBody) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  processing_mode:
-    response_header_mode: "SEND"
-    response_body_mode: "FULL_DUPLEX_STREAMED"
-    response_trailer_mode: "SEND"
-  )EOF");
-
-  // Create synthetic HTTP request
-  HttpTestUtility::addDefaultHeaders(request_headers_);
-  request_headers_.setMethod("POST");
-  request_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->decodeHeaders(request_headers_, false));
-  processRequestHeaders(false, absl::nullopt);
-
-  response_headers_.addCopy(LowerCaseString(":status"), "200");
-  response_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  // Envoy sends header, body and trailer.
-  bool encoding_watermarked = false;
-  setUpEncodingWatermarking(encoding_watermarked);
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->encodeHeaders(response_headers_, false));
-  EXPECT_EQ(FilterTrailersStatus::StopIteration, filter_->encodeTrailers(response_trailers_));
-
-  // Server now sends back response.
-  processResponseHeadersAfterTrailer(absl::nullopt);
-  processResponseTrailers(absl::nullopt, true);
-
-  EXPECT_EQ(config_->stats().spurious_msgs_received_.value(), 0);
-  filter_->onDestroy();
-}
-
-TEST_F(HttpFilterTest, DuplexStreamedBodyProcessingTestWithFilterConfigMissing) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  processing_mode:
-    response_body_mode: "STREAMED"
-  )EOF");
-
-  // Create synthetic HTTP request
-  HttpTestUtility::addDefaultHeaders(request_headers_);
-  request_headers_.setMethod("POST");
-  request_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->decodeHeaders(request_headers_, false));
-  processRequestHeaders(false, absl::nullopt);
-
-  response_headers_.addCopy(LowerCaseString(":status"), "200");
-  response_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  bool encoding_watermarked = false;
-  setUpEncodingWatermarking(encoding_watermarked);
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->encodeHeaders(response_headers_, false));
-  processResponseHeaders(false, absl::nullopt);
-
-  for (int i = 0; i < 4; i++) {
-    // 4 request chunks are sent to the ext_proc server.
-    Buffer::OwnedImpl resp_chunk;
-    TestUtility::feedBufferWithRandomCharacters(resp_chunk, 100);
-    EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(resp_chunk, false));
-  }
-
-  processResponseBody(
-      [](const HttpBody&, ProcessingResponse&, BodyResponse& resp) {
-        auto* streamed_response =
-            resp.mutable_response()->mutable_body_mutation()->mutable_streamed_response();
-        streamed_response->set_body("AAA");
-      },
-      false);
-
-  // Verify spurious message is received.
-  EXPECT_EQ(config_->stats().spurious_msgs_received_.value(), 1);
-  filter_->onDestroy();
-}
-
-TEST_F(HttpFilterTest, SendNormalBodyMutationTestWithFilterConfigDuplexStreamed) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  failure_mode_allow: true
-  processing_mode:
-    response_body_mode: "FULL_DUPLEX_STREAMED"
-    response_trailer_mode: "SEND"
-  )EOF");
-
-  // Create synthetic HTTP request
-  HttpTestUtility::addDefaultHeaders(request_headers_);
-  request_headers_.setMethod("POST");
-  request_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->decodeHeaders(request_headers_, false));
-  processRequestHeaders(false, absl::nullopt);
-
-  response_headers_.addCopy(LowerCaseString(":status"), "200");
-  response_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-
-  bool encoding_watermarked = false;
-  setUpEncodingWatermarking(encoding_watermarked);
-  EXPECT_EQ(FilterHeadersStatus::StopIteration, filter_->encodeHeaders(response_headers_, false));
-  processResponseHeaders(false, absl::nullopt);
-
-  Buffer::OwnedImpl resp_chunk;
-  TestUtility::feedBufferWithRandomCharacters(resp_chunk, 100);
-  EXPECT_EQ(FilterDataStatus::StopIterationNoBuffer, filter_->encodeData(resp_chunk, true));
-
-  processResponseBody(
-      [](const HttpBody&, ProcessingResponse&, BodyResponse& resp) {
-        auto* body_mut = resp.mutable_response()->mutable_body_mutation();
-        body_mut->set_body("AAA");
-      },
-      true);
-
-  // Verify spurious message is received.
-  EXPECT_EQ(config_->stats().spurious_msgs_received_.value(), 1);
-  filter_->onDestroy();
-}
-
 // Verify if ext_proc filter is in the upstream filter chain, and if the ext_proc server
 // sends back response with clear_route_cache set to true, it is ignored.
 TEST_F(HttpFilterTest, ClearRouteCacheHeaderMutationUpstreamIgnored) {
@@ -5110,419 +3943,6 @@ TEST_F(HttpFilterTest, PostAndRespondImmediatelyUpstream) {
   EXPECT_EQ(config_->stats().stream_msgs_sent_.value(), 1);
   EXPECT_EQ(config_->stats().stream_msgs_received_.value(), 1);
   EXPECT_EQ(config_->stats().streams_closed_.value(), 1);
-}
-
-TEST_F(HttpFilterTest, HeaderProcessingInObservabilityMode) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  observability_mode: true
-  )EOF");
-
-  EXPECT_TRUE(config_->observabilityMode());
-  observability_mode_ = true;
-
-  // Create synthetic HTTP request
-  request_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-  request_headers_.addCopy(LowerCaseString("content-length"), 10);
-  request_headers_.addCopy(LowerCaseString("x-some-other-header"), "yes");
-
-  // In the observability mode, the filter returns `Continue` in all events of http request
-  // lifecycle.
-  EXPECT_EQ(FilterHeadersStatus::Continue, filter_->decodeHeaders(request_headers_, false));
-  processRequestHeaders(false,
-                        [](const HttpHeaders& header_req, ProcessingResponse&, HeadersResponse&) {
-                          EXPECT_FALSE(header_req.end_of_stream());
-                          TestRequestHeaderMapImpl expected{{":path", "/"},
-                                                            {":method", "POST"},
-                                                            {":scheme", "http"},
-                                                            {"host", "host"},
-                                                            {"content-type", "text/plain"},
-                                                            {"content-length", "10"},
-                                                            {"x-some-other-header", "yes"}};
-                          EXPECT_THAT(header_req.headers(), HeaderProtosEqual(expected));
-                        });
-
-  Buffer::OwnedImpl req_data("foo");
-  EXPECT_EQ(FilterDataStatus::Continue, filter_->decodeData(req_data, true));
-  EXPECT_EQ(FilterTrailersStatus::Continue, filter_->decodeTrailers(request_trailers_));
-
-  response_headers_.addCopy(LowerCaseString(":status"), "200");
-  response_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-  response_headers_.addCopy(LowerCaseString("content-length"), "3");
-
-  // In the observability mode, the filter returns `Continue` in all events of http response
-  // lifecycle.
-  EXPECT_EQ(FilterHeadersStatus::Continue, filter_->encodeHeaders(response_headers_, false));
-  processResponseHeaders(
-      false, [](const HttpHeaders& header_resp, ProcessingResponse&, HeadersResponse&) {
-        EXPECT_FALSE(header_resp.end_of_stream());
-        TestRequestHeaderMapImpl expected_response{
-            {":status", "200"}, {"content-type", "text/plain"}, {"content-length", "3"}};
-        EXPECT_THAT(header_resp.headers(), HeaderProtosEqual(expected_response));
-      });
-
-  Buffer::OwnedImpl resp_data("foo");
-  EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(resp_data, false));
-  Buffer::OwnedImpl empty_data;
-  EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(empty_data, true));
-  EXPECT_EQ(FilterTrailersStatus::Continue, filter_->encodeTrailers(response_trailers_));
-
-  deferred_close_timer_ = new Event::MockTimer(&dispatcher_);
-  // Deferred close timer is expected to be enabled by `DeferredDeletableStream`'s deferredClose(),
-  // which is triggered by filter onDestroy() function below.
-  EXPECT_CALL(*deferred_close_timer_,
-              enableTimer(std::chrono::milliseconds(DEFAULT_DEFERRED_CLOSE_TIMEOUT_MS), _));
-  filter_->onDestroy();
-  deferred_close_timer_->invokeCallback();
-
-  EXPECT_EQ(1, config_->stats().streams_started_.value());
-  // Two messages (request and response header message) are sent.
-  EXPECT_EQ(2, config_->stats().stream_msgs_sent_.value());
-  // No response is received in observability mode.
-  EXPECT_EQ(0, config_->stats().stream_msgs_received_.value());
-  // Deferred stream is closed.
-  EXPECT_EQ(1, config_->stats().streams_closed_.value());
-}
-
-TEST_F(HttpFilterTest, StreamingBodiesInObservabilityMode) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  observability_mode: true
-  processing_mode:
-    request_body_mode: "STREAMED"
-    response_body_mode: "STREAMED"
-  )EOF");
-
-  uint32_t content_length = 100;
-  observability_mode_ = true;
-
-  // Create synthetic HTTP request
-  HttpTestUtility::addDefaultHeaders(request_headers_);
-  request_headers_.setMethod("POST");
-  request_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-  request_headers_.addCopy(LowerCaseString("content-length"), content_length);
-
-  EXPECT_CALL(decoder_callbacks_, decodingBuffer()).WillRepeatedly(Return(nullptr));
-  EXPECT_EQ(FilterHeadersStatus::Continue, filter_->decodeHeaders(request_headers_, false));
-  processRequestHeaders(false, absl::nullopt);
-  // In observability mode, content length is not removed as there is no mutation from ext_proc
-  // server.
-  EXPECT_EQ(request_headers_.getContentLengthValue(), absl::StrCat(content_length));
-
-  Buffer::OwnedImpl want_request_body;
-  Buffer::OwnedImpl got_request_body;
-  EXPECT_CALL(decoder_callbacks_, injectDecodedDataToFilterChain(_, true))
-      .WillRepeatedly(Invoke(
-          [&got_request_body](Buffer::Instance& data, Unused) { got_request_body.move(data); }));
-
-  Buffer::OwnedImpl req_chunk_1;
-  TestUtility::feedBufferWithRandomCharacters(req_chunk_1, 100);
-  want_request_body.add(req_chunk_1.toString());
-  EXPECT_EQ(FilterDataStatus::Continue, filter_->decodeData(req_chunk_1, true));
-  got_request_body.move(req_chunk_1);
-  EXPECT_EQ(want_request_body.toString(), got_request_body.toString());
-
-  response_headers_.addCopy(LowerCaseString(":status"), "200");
-  response_headers_.addCopy(LowerCaseString("content-type"), "text/plain");
-  response_headers_.addCopy(LowerCaseString("content-length"), content_length);
-
-  EXPECT_CALL(encoder_callbacks_, encodingBuffer()).WillRepeatedly(Return(nullptr));
-  EXPECT_EQ(FilterHeadersStatus::Continue, filter_->encodeHeaders(response_headers_, false));
-  processResponseHeaders(false, absl::nullopt);
-  EXPECT_EQ(response_headers_.getContentLengthValue(), absl::StrCat(content_length));
-
-  Buffer::OwnedImpl want_response_body;
-  Buffer::OwnedImpl got_response_body;
-  EXPECT_CALL(encoder_callbacks_, injectEncodedDataToFilterChain(_, _))
-      .WillRepeatedly(Invoke(
-          [&got_response_body](Buffer::Instance& data, Unused) { got_response_body.move(data); }));
-
-  for (int i = 0; i < 5; i++) {
-    Buffer::OwnedImpl resp_chunk;
-    TestUtility::feedBufferWithRandomCharacters(resp_chunk, 100);
-    want_response_body.add(resp_chunk.toString());
-    EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(resp_chunk, false));
-    got_response_body.move(resp_chunk);
-  }
-
-  Buffer::OwnedImpl last_resp_chunk;
-  EXPECT_EQ(FilterDataStatus::Continue, filter_->encodeData(last_resp_chunk, true));
-
-  // At this point, since we injected the data from each chunk after the "encodeData"
-  // callback, and since we also injected any chunks inserted using "injectEncodedData,"
-  // the two buffers should match!
-  EXPECT_EQ(want_response_body.toString(), got_response_body.toString());
-
-  deferred_close_timer_ = new Event::MockTimer(&dispatcher_);
-  // Deferred close timer is expected to be enabled by `DeferredDeletableStream`'s deferredClose(),
-  // which is triggered by filter onDestroy() function.
-  EXPECT_CALL(*deferred_close_timer_,
-              enableTimer(std::chrono::milliseconds(DEFAULT_DEFERRED_CLOSE_TIMEOUT_MS), _));
-  filter_->onDestroy();
-  deferred_close_timer_->invokeCallback();
-
-  EXPECT_EQ(1, config_->stats().streams_started_.value());
-  EXPECT_EQ(9, config_->stats().stream_msgs_sent_.value());
-  EXPECT_EQ(0, config_->stats().stream_msgs_received_.value());
-  EXPECT_EQ(1, config_->stats().streams_closed_.value());
-}
-
-TEST_F(HttpFilterTest, StreamingAllDataInObservabilityMode) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  observability_mode: true
-  processing_mode:
-    request_header_mode: "SEND"
-    response_header_mode: "SEND"
-    request_body_mode: "STREAMED"
-    response_body_mode: "STREAMED"
-    request_trailer_mode: "SEND"
-    response_trailer_mode: "SEND"
-  )EOF");
-
-  observability_mode_ = true;
-
-  HttpTestUtility::addDefaultHeaders(request_headers_);
-  request_headers_.setMethod("POST");
-  EXPECT_CALL(decoder_callbacks_, decodingBuffer()).WillRepeatedly(Return(nullptr));
-  EXPECT_EQ(FilterHeadersStatus::Continue, filter_->decodeHeaders(request_headers_, false));
-  processRequestHeaders(false, absl::nullopt);
-
-  const uint32_t chunk_number = 20;
-  sendChunkRequestData(chunk_number, true);
-  EXPECT_EQ(FilterTrailersStatus::Continue, filter_->decodeTrailers(request_trailers_));
-  processRequestTrailers(absl::nullopt);
-
-  response_headers_.addCopy(LowerCaseString(":status"), "200");
-  EXPECT_EQ(FilterHeadersStatus::Continue, filter_->encodeHeaders(response_headers_, false));
-  processResponseHeaders(true, absl::nullopt);
-  sendChunkResponseData(chunk_number * 2, true);
-  EXPECT_EQ(FilterTrailersStatus::Continue, filter_->encodeTrailers(response_trailers_));
-  processResponseTrailers(absl::nullopt, false);
-  deferred_close_timer_ = new Event::MockTimer(&dispatcher_);
-  // Deferred close timer is expected to be enabled by `DeferredDeletableStream`'s deferredClose(),
-  // which is triggered by filter onDestroy() function.
-  EXPECT_CALL(*deferred_close_timer_,
-              enableTimer(std::chrono::milliseconds(DEFAULT_DEFERRED_CLOSE_TIMEOUT_MS), _));
-  filter_->onDestroy();
-  deferred_close_timer_->invokeCallback();
-
-  EXPECT_EQ(1, config_->stats().streams_started_.value());
-  // Total gRPC messages include two headers and two trailers on top of the req/resp chunk data.
-  uint32_t total_msg = 3 * chunk_number + 4;
-  EXPECT_EQ(total_msg, config_->stats().stream_msgs_sent_.value());
-  EXPECT_EQ(0, config_->stats().stream_msgs_received_.value());
-  EXPECT_EQ(1, config_->stats().streams_closed_.value());
-}
-
-class HttpFilter2Test : public HttpFilterTest,
-                        public ::Envoy::Http::HttpConnectionManagerImplMixin {};
-
-// Test proves that when decodeData(data, end_stream=true) is called before request headers response
-// is returned, ext_proc filter will buffer the data in the ActiveStream buffer without triggering a
-// buffer over high watermark call, which ends in an 413 error return on request path.
-TEST_F(HttpFilter2Test, LastDecodeDataCallExceedsStreamBufferLimitWouldJustRaiseHighWatermark) {
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  )EOF");
-  HttpConnectionManagerImplMixin::setup(Envoy::Http::SetupOpts().setServerName("fake-server"));
-  HttpConnectionManagerImplMixin::initial_buffer_limit_ = 10;
-  HttpConnectionManagerImplMixin::setUpBufferLimits();
-
-  std::shared_ptr<MockStreamDecoderFilter> mock_filter(new NiceMock<MockStreamDecoderFilter>());
-  EXPECT_CALL(filter_factory_, createFilterChain(_))
-      .WillOnce(Invoke([&](::Envoy::Http::FilterChainManager& manager) -> bool {
-        // Add ext_proc filter.
-        FilterFactoryCb cb = [&](FilterChainFactoryCallbacks& callbacks) {
-          callbacks.addStreamDecoderFilter(filter_);
-        };
-        manager.applyFilterFactoryCb({}, cb);
-        // Add the mock-decoder filter.
-        FilterFactoryCb mock_filter_cb = [&](FilterChainFactoryCallbacks& callbacks) {
-          callbacks.addStreamDecoderFilter(mock_filter);
-        };
-        manager.applyFilterFactoryCb({}, mock_filter_cb);
-
-        return true;
-      }));
-  EXPECT_CALL(*mock_filter, decodeHeaders(_, false))
-      .WillOnce(Invoke([&](RequestHeaderMap& headers, bool end_stream) {
-        // The next decoder filter should be able to see the mutations made by the external server.
-        EXPECT_FALSE(end_stream);
-        EXPECT_EQ(headers.Path()->value().getStringView(), "/mutated_path/bluh");
-        EXPECT_EQ(headers.get(Envoy::Http::LowerCaseString("foo"))[0]->value().getStringView(),
-                  "gift-from-external-server");
-        mock_filter->callbacks_->sendLocalReply(::Envoy::Http::Code::OK,
-                                                "Direct response from mock filter.", nullptr,
-                                                absl::nullopt, "");
-        return FilterHeadersStatus::StopIteration;
-      }));
-  EXPECT_CALL(response_encoder_, encodeHeaders(_, _))
-      .WillOnce(Invoke([&](const ResponseHeaderMap& headers, bool end_stream) -> void {
-        EXPECT_FALSE(end_stream);
-        EXPECT_EQ(headers.Status()->value().getStringView(), "200");
-      }));
-  EXPECT_CALL(response_encoder_, encodeData(_, _))
-      .WillOnce(Invoke([&](Buffer::Instance& data, bool end_stream) -> void {
-        EXPECT_TRUE(end_stream);
-        EXPECT_EQ(data.toString(), "Direct response from mock filter.");
-      }));
-  // Start the request.
-  EXPECT_CALL(*codec_, dispatch(_))
-      .WillOnce(Invoke([&](Buffer::Instance& data) -> ::Envoy::Http::Status {
-        EXPECT_EQ(data.length(), 5);
-        data.drain(5);
-
-        HttpConnectionManagerImplMixin::decoder_ = &conn_manager_->newStream(response_encoder_);
-        RequestHeaderMapPtr headers{new TestRequestHeaderMapImpl{
-            {":authority", "host"}, {":path", "/bluh"}, {":method", "GET"}}};
-        HttpConnectionManagerImplMixin::decoder_->decodeHeaders(std::move(headers), false);
-        Buffer::OwnedImpl request_body("Definitely more than 10 bytes data.");
-        HttpConnectionManagerImplMixin::decoder_->decodeData(request_body, true);
-        // Now external server returns the request header response.
-        auto response = std::make_unique<ProcessingResponse>();
-        auto* headers_response = response->mutable_request_headers();
-        auto* hdr =
-            headers_response->mutable_response()->mutable_header_mutation()->add_set_headers();
-        hdr->mutable_append()->set_value(false);
-        hdr->mutable_header()->set_key("foo");
-        hdr->mutable_header()->set_raw_value("gift-from-external-server");
-        hdr = headers_response->mutable_response()->mutable_header_mutation()->add_set_headers();
-        hdr->mutable_append()->set_value(false);
-        hdr->mutable_header()->set_key(":path");
-        hdr->mutable_header()->set_raw_value("/mutated_path/bluh");
-        HttpFilterTest::stream_callbacks_->onReceiveMessage(std::move(response));
-
-        return ::Envoy::Http::okStatus();
-      }));
-
-  Buffer::OwnedImpl fake_input("hello");
-  conn_manager_->onData(fake_input, false);
-}
-
-// Test proves that when encodeData(data, end_stream=true) is called before headers response is
-// returned, ext_proc filter will buffer the data in the ActiveStream buffer without triggering a
-// buffer over high watermark call, which ends in a 500 error on response path.
-TEST_F(HttpFilter2Test, LastEncodeDataCallExceedsStreamBufferLimitWouldJustRaiseHighWatermark) {
-  // Configure the filter to only pass response headers to ext server.
-  initialize(R"EOF(
-  grpc_service:
-    envoy_grpc:
-      cluster_name: "ext_proc_server"
-  processing_mode:
-    request_header_mode: "SKIP"
-    response_header_mode: "SEND"
-    request_body_mode: "NONE"
-    response_body_mode: "NONE"
-    request_trailer_mode: "SKIP"
-    response_trailer_mode: "SKIP"
-
-  )EOF");
-  HttpConnectionManagerImplMixin::setup(Envoy::Http::SetupOpts().setServerName("fake-server"));
-  HttpConnectionManagerImplMixin::initial_buffer_limit_ = 10;
-  HttpConnectionManagerImplMixin::setUpBufferLimits();
-
-  std::shared_ptr<MockStreamEncoderFilter> mock_encode_filter(
-      new NiceMock<MockStreamEncoderFilter>());
-  std::shared_ptr<MockStreamDecoderFilter> mock_decode_filter(
-      new NiceMock<MockStreamDecoderFilter>());
-
-  EXPECT_CALL(*mock_encode_filter, encodeHeaders(_, _))
-      .WillOnce(Invoke([&](ResponseHeaderMap& headers, bool end_stream) {
-        EXPECT_FALSE(end_stream);
-        // The last encode filter will see the mutations from ext server.
-        // NOTE: Without raising a high watermark when end_stream is true in onData(), if the stream
-        // buffer high watermark reached, a 500 response too large error is raised.
-        EXPECT_EQ(headers.Status()->value().getStringView(), "200");
-        EXPECT_EQ(headers.get(Envoy::Http::LowerCaseString("foo"))[0]->value().getStringView(),
-                  "gift-from-external-server");
-        EXPECT_EQ(headers.get(Envoy::Http::LowerCaseString("new_response_header"))[0]
-                      ->value()
-                      .getStringView(),
-                  "bluh");
-
-        return FilterHeadersStatus::Continue;
-      }));
-  EXPECT_CALL(*mock_encode_filter, encodeData(_, true))
-      .WillOnce(Invoke([&](Buffer::Instance& data, bool end_stream) {
-        EXPECT_TRUE(end_stream);
-        EXPECT_EQ(data.toString(),
-                  "Direct response from mock filter, Definitely more than 10 bytes data.");
-        return FilterDataStatus::Continue;
-      }));
-  EXPECT_CALL(filter_factory_, createFilterChain(_))
-      .WillOnce(Invoke([&](::Envoy::Http::FilterChainManager& manager) -> bool {
-        // Add the mock-encoder filter.
-        FilterFactoryCb mock_encode_filter_cb = [&](FilterChainFactoryCallbacks& callbacks) {
-          callbacks.addStreamEncoderFilter(mock_encode_filter);
-        };
-        manager.applyFilterFactoryCb({}, mock_encode_filter_cb);
-
-        // Add ext_proc filter.
-        FilterFactoryCb cb = [&](FilterChainFactoryCallbacks& callbacks) {
-          callbacks.addStreamFilter(filter_);
-        };
-        manager.applyFilterFactoryCb({}, cb);
-        // Add the mock-decoder filter.
-        FilterFactoryCb mock_decode_filter_cb = [&](FilterChainFactoryCallbacks& callbacks) {
-          callbacks.addStreamDecoderFilter(mock_decode_filter);
-        };
-        manager.applyFilterFactoryCb({}, mock_decode_filter_cb);
-
-        return true;
-      }));
-  EXPECT_CALL(*mock_decode_filter, decodeHeaders(_, _))
-      .WillOnce(Invoke([&](RequestHeaderMap& headers, bool end_stream) {
-        EXPECT_TRUE(end_stream);
-        EXPECT_EQ(headers.Path()->value().getStringView(), "/bluh");
-        // Direct response from decode filter.
-        ResponseHeaderMapPtr response_headers{
-            new TestResponseHeaderMapImpl{{":status", "200"}, {"foo", "foo-value"}}};
-        mock_decode_filter->callbacks_->encodeHeaders(std::move(response_headers), false,
-                                                      "filter_direct_response");
-        // Send a large body in one shot.
-        Buffer::OwnedImpl fake_response(
-            "Direct response from mock filter, Definitely more than 10 bytes data.");
-        mock_decode_filter->callbacks_->encodeData(fake_response, true);
-
-        // Now return from ext server the response for processing response headers.
-        EXPECT_TRUE(last_request_.has_response_headers());
-        auto response = std::make_unique<ProcessingResponse>();
-        auto* headers_response = response->mutable_response_headers();
-        auto* hdr =
-            headers_response->mutable_response()->mutable_header_mutation()->add_set_headers();
-        hdr->mutable_append()->set_value(false);
-        hdr->mutable_header()->set_key("foo");
-        hdr->mutable_header()->set_raw_value("gift-from-external-server");
-        hdr = headers_response->mutable_response()->mutable_header_mutation()->add_set_headers();
-        hdr->mutable_append()->set_value(false);
-        hdr->mutable_header()->set_key("new_response_header");
-        hdr->mutable_header()->set_raw_value("bluh");
-        HttpFilterTest::stream_callbacks_->onReceiveMessage(std::move(response));
-        return FilterHeadersStatus::StopIteration;
-      }));
-  // Start the request.
-  EXPECT_CALL(*codec_, dispatch(_))
-      .WillOnce(Invoke([&](Buffer::Instance& data) -> ::Envoy::Http::Status {
-        EXPECT_EQ(data.length(), 5);
-        data.drain(5);
-        HttpConnectionManagerImplMixin::decoder_ = &conn_manager_->newStream(response_encoder_);
-        RequestHeaderMapPtr headers{new TestRequestHeaderMapImpl{
-            {":authority", "host"}, {":path", "/bluh"}, {":method", "GET"}}};
-        HttpConnectionManagerImplMixin::decoder_->decodeHeaders(std::move(headers), true);
-        return ::Envoy::Http::okStatus();
-      }));
-
-  Buffer::OwnedImpl fake_input("hello");
-  conn_manager_->onData(fake_input, false);
 }
 
 // Test that per route metadata override does override inherited grpc_service configuration.
